+++ AccountManagement.Domain/RoleAgg/Permission.cs
@@ -3,12 +3,12 @@
     public class Permission
     {
         public long Id { get; private set; }
-        public long Code { get; private set; }
+        public int Code { get; private set; }
         public string Name { get; private set; }
         public long RoleId { get; private set; }
         public Role Role { get; private set; }
 
-        public Permission(long code)
+        public Permission(int code)
         {
             Code = code;
         }
@@ -20,3 +20,4 @@
         }
     }
 }
+
